+++ api/src/main/java/org/allaymc/api/item/data/TrimPattern.java
@@ -1,0 +1,9 @@
+package org.allaymc.api.item.data;
+
+import org.allaymc.api.item.type.ItemType;
+
+/**
+ * @author daoge_cmd
+ */
+public record TrimPattern(ItemType<?> itemType, String patternId) {
+}